refactor: 💡 named backfill renderer and a dedicated org-context format module
The card re-render becomes one module-private async function with a
real try/catch instead of a then/catch chain, and the compact
plan/feature renderers move out of the service into
orgContextFormat.ts so the service holds only load and cache.
