+++ apps/leaf/src/internal/approvals/surfaces/slack/present.ts
@@ -49,6 +49,61 @@ export const postApprovalCardForRow = async ({
 	}
 };
 
+/** Grouped step previews land after the card is already visible; once they
+ * persist, the card re-renders — unless the approval resolved meanwhile, whose
+ * card must not be overwritten. */
+const renderBackfilledGroupCard = async ({
+	backfill,
+	channelId,
+	created,
+	env,
+	logger,
+	messageId,
+	target,
+}: {
+	backfill: () => Promise<Record<string, unknown> | undefined>;
+	channelId: string;
+	created: { approvalId: string; preview: unknown; toolName: string };
+	env: ChatApproval["env"];
+	logger: AutumnLogger;
+	messageId: string;
+	target: ReplyTarget;
+}) => {
+	try {
+		const enrichedToolArgs = await backfill();
+		if (!enrichedToolArgs) return;
+		if (!target.adapter?.editMessage) {
+			logger.warn("No adapter to re-render the backfilled card", {
+				event: "leaf.approval_group_preview_render_skipped",
+				approval_id: created.approvalId,
+			});
+			return;
+		}
+		const current = await chatApprovalRepo.get({
+			approvalId: created.approvalId,
+			db,
+		});
+		if (current?.status !== "pending") return;
+		await target.adapter.editMessage(
+			channelId,
+			messageId,
+			approvalCard({
+				id: created.approvalId,
+				env,
+				preview: created.preview,
+				toolArgs: enrichedToolArgs,
+				toolName: created.toolName,
+			}),
+		);
+	} catch (error) {
+		logger.warn("Could not backfill grouped previews", {
+			event: "leaf.approval_group_preview_backfill_failed",
+			approval_id: created.approvalId,
+			error,
+		});
+	}
+};
+
 export const presentApproval = async ({
 	channelId,
 	installation,
@@ -108,46 +163,16 @@ export const presentApproval = async ({
 			error,
 		});
 	}
-	// Grouped step previews land after the card is already visible; the row
-	// update is pending-guarded so a card resolved meanwhile is left alone.
 	if (created.backfillGroupedPreviews) {
-		void created
-			.backfillGroupedPreviews()
-			.then(async (enrichedToolArgs) => {
-				if (!enrichedToolArgs) return;
-				if (!target.adapter?.editMessage) {
-					logger.warn("No adapter to re-render the backfilled card", {
-						event: "leaf.approval_group_preview_render_skipped",
-						approval_id: created.approvalId,
-					});
-					return;
-				}
-				// The row update is pending-guarded, but the Slack edit races the
-				// approve click — re-check before overwriting a resolved card.
-				const current = await chatApprovalRepo.get({
-					approvalId: created.approvalId,
-					db,
-				});
-				if (current?.status !== "pending") return;
-				await target.adapter.editMessage(
-					channelId,
-					sent.id,
-					approvalCard({
-						id: created.approvalId,
-						env,
-						preview: created.preview,
-						toolArgs: enrichedToolArgs,
-						toolName: created.toolName,
-					}),
-				);
-			})
-			.catch((error) => {
-				logger.warn("Could not backfill grouped previews", {
-					event: "leaf.approval_group_preview_backfill_failed",
-					approval_id: created.approvalId,
-					error,
-				});
-			});
+		void renderBackfilledGroupCard({
+			backfill: created.backfillGroupedPreviews,
+			channelId,
+			created,
+			env,
+			logger,
+			messageId: sent.id,
+			target,
+		});
 	}
 	return true;
 };
