+++ apps/leaf/src/internal/autumnMcp/orgContextService.ts
@@ -1,8 +1,12 @@
 import type { AutumnLogger } from "@autumn/logging";
-import { asRecord, parsePreviewPayload } from "@autumn/render";
 import { type AppEnv, ms } from "@autumn/shared";
 import { createTtlCache } from "../../lib/ttlCache.js";
 import { executeAutumnMcpTool } from "./client.js";
+import {
+	compactFeatures,
+	compactPlans,
+	toJsonBlock,
+} from "./orgContextFormat.js";
 
 export type AutumnOrgContext = {
 	instructions?: string;
@@ -10,74 +14,6 @@ export type AutumnOrgContext = {
 };
 
 type ExecuteAutumnTool = typeof executeAutumnMcpTool;
-
-const toJsonBlock = ({
-	label,
-	note,
-	pretty = true,
-	value,
-}: {
-	label: string;
-	note?: string;
-	pretty?: boolean;
-	value: unknown;
-}) =>
-	`${label}${note ? ` (${note})` : ""}:\n\`\`\`json\n${JSON.stringify(value, null, pretty ? 2 : undefined)}\n\`\`\``;
-
-const listOf = (value: unknown): Record<string, unknown>[] => {
-	const unwrapped = Array.isArray(value)
-		? value
-		: (parsePreviewPayload(value) ?? value);
-	const record = asRecord(unwrapped) ?? {};
-	const list = [record.list, record.plans, record.features, unwrapped].find(
-		Array.isArray,
-	);
-	return (list ?? []).map((entry) => asRecord(entry) ?? {});
-};
-
-const compactPrice = (price: unknown) => {
-	const record = asRecord(price) ?? {};
-	if (record.amount === undefined) return undefined;
-	return `${record.amount}/${record.interval ?? "one_off"}`;
-};
-
-const compactItem = (item: unknown) => {
-	const record = asRecord(item) ?? {};
-	const featureId = record.feature_id ?? record.id;
-	if (typeof featureId !== "string") return undefined;
-	const parts = [featureId];
-	if (record.included !== undefined && record.included !== null) {
-		parts.push(`included=${record.included}`);
-	}
-	const price = asRecord(record.price) ?? {};
-	if (price.billing_method) parts.push(String(price.billing_method));
-	return parts.join(" ");
-};
-
-/** The 30KB pretty-printed listPlans/listFeatures dump is ~89% whitespace and
- * display noise; the orchestrator only routes and answers trivial questions,
- * so it gets a compact index and fetches details with its own tools. */
-const compactPlans = (plans: unknown) =>
-	listOf(plans).map((plan) => ({
-		...(plan.add_on === true ? { add_on: true } : {}),
-		id: plan.id,
-		...(Array.isArray(plan.items) && plan.items.length
-			? {
-					items: plan.items
-						.map(compactItem)
-						.filter((item): item is string => Boolean(item)),
-				}
-			: {}),
-		name: plan.name,
-		...(compactPrice(plan.price) ? { price: compactPrice(plan.price) } : {}),
-	}));
-
-const compactFeatures = (features: unknown) =>
-	listOf(features).map((feature) => ({
-		id: feature.id,
-		name: feature.name,
-		...(feature.type ? { type: feature.type } : {}),
-	}));
 
 const withoutNotes = (value: unknown) => {
 	if (!value || typeof value !== "object" || Array.isArray(value)) return value;
